Remove unused endpoints & associated tests (#2)
Removes unused continuous set and RNAQuant/expression endpoints + tests, and thus the need for the PyBigWig dependency.
