+++ .travis.yml
@@ -38,22 +38,8 @@ script:
            tests ga4gh scripts
   - python -mnose tests 
               --with-coverage --cover-package ga4gh.server
-              --cover-inclusive --cover-min-percentage 80
+              --cover-inclusive --cover-min-percentage 70
               --cover-branches --cover-erase
-  - make clean -C docs
-  - make -C docs
 
-after_success:
-# run codecov
-  - bash <(curl -s https://codecov.io/bash)
-
-# deploy to pypi when there is a new tag in master after a successful build
-deploy:
-    provider: pypi
-    user: dannycolligan
-    password:
-        secure: ffTNL0M96N0j4rLNxnBXUFg8R+7r0FJG2yXOsog/9QIK5NNIKrgI/a0CfIMxR4TCdTkK3x2O+k6gwrjiNmGAVi93brOIJHV5D5nq5PM5P0TnugIHVY6iwufltvWZn1NN2/8n93RKjNXya77nR53sUhamIGLad8LK8eQGlOU5EMg=
-    skip_upload_docs: true
-    on:
-        repo: ga4gh/ga4gh-server
-        tags: true
+#   - make clean -C docs
+#   - make -C docs
