+++ tests/datadriven/test_continuous.py
@@ -19,6 +19,7 @@ import ga4gh.server.exceptions as exceptions
 import tests.paths as paths
 
 
+@unittest.skip("Disabled")
 class TestContinuous(unittest.TestCase):
     """
     Unit tests for continuous data
