+++ tests/datadriven/test_rna_quantification.py
@@ -20,6 +20,8 @@ import tests.paths as paths
 
 import ga4gh.schemas.protocol as protocol
 
+import unittest
+
 
 _datasetName = "ds"
 
@@ -81,6 +83,7 @@ def testRnaQuantification():
         yield test
 
 
+@unittest.skip("Disabled")
 class RnaQuantificationTest(datadriven.DataDrivenTest):
     """
     Data driven test class for rna quantification. Builds an alternative model
