+++ tests/datadriven/test_sequence_annotations.py
@@ -15,6 +15,7 @@ import tests.paths as paths
 
 import ga4gh.schemas.protocol as protocol
 
+
 _datasetName = "ds"
 
 _discontinuousTestData = {
