+++ tests/end_to_end/test_sequence_annotations.py
@@ -164,6 +164,7 @@ class TestSequenceAnnotations(unittest.TestCase):
             path, request, protocol.SearchContinuousSetsResponse)
         return responseData.continuous_sets
 
+    @unittest.skip("Disabled")
     def testSearchContinuous(self):
         continuousSets = self.getAllContinuousSets()
         for continuousSet in continuousSets:
