+++ tests/unit/test_client.py
@@ -249,6 +249,7 @@ class ExhaustiveListingsMixin(object):
                         for dmRead, read in utils.zipLists(dmReads, reads):
                             self.assertEqual(dmRead, read)
 
+    @unittest.skip("Disabled")
     def testAllRnaQuantificationSets(self):
         for dataset in self.client.search_datasets():
             rnaQuantificationSets = \
