+++ tests/unit/test_data_interface.py
@@ -123,6 +123,7 @@ class TestInterfacingLayer(unittest.TestCase):
             feature = self._client.get_feature(featureId)
             self.assertEqual(repoFeature, feature)
 
+    @unittest.skip("Disabled")
     def testGetContinuousSet(self):
         self._testGetMethod(
             self._repo.allContinuousSets, self._client.get_continuous_set)
@@ -135,16 +136,19 @@ class TestInterfacingLayer(unittest.TestCase):
         self._testGetMethod(
             self._repo.allReferences, self._client.get_reference)
 
+    @unittest.skip("Disabled")
     def testGetRnaQuantificationSet(self):
         self._testGetMethod(
             self._repo.allRnaQuantificationSets,
             self._client.get_rna_quantification_set)
 
+    @unittest.skip("Disabled")
     def testGetRnaQuantification(self):
         self._testGetMethod(
             self._repo.allRnaQuantifications,
             self._client.get_rna_quantification)
 
+    @unittest.skip("Disabled")
     def testGetExpressionLevel(self):
         self._testGetMethod(
             self._repo.allExpressionLevels,
@@ -220,6 +224,7 @@ class TestInterfacingLayer(unittest.TestCase):
             self._client.search_feature_sets,
             self._repo.getDatasets())
 
+    @unittest.skip("Disabled")
     def testSearchContinuousSets(self):
         self._testSearchMethodInContainer(
             'getContinuousSets',
@@ -319,18 +324,21 @@ class TestInterfacingLayer(unittest.TestCase):
             self._repo.allPhenotypeAssociationSets(),
             equalMethod='assertEqual')
 
+    @unittest.skip("Disabled")
     def testSearchRnaQuantificationSets(self):
         self._testSearchMethodInContainer(
             'getRnaQuantificationSets',
             self._client.search_rna_quantification_sets,
             self._repo.getDatasets())
 
+    @unittest.skip("Disabled")
     def testSearchRnaQuantifications(self):
         self._testSearchMethodInContainer(
             'getRnaQuantifications',
             self._client.search_rna_quantifications,
             self._repo.allRnaQuantificationSets())
 
+    @unittest.skip("Disabled")
     def testSearchExpressionLevels(self):
         self._testSearchMethodInContainer(
             'getExpressionLevels',
