+++ tests/unit/test_views.py
@@ -72,14 +72,14 @@ class TestFrontend(unittest.TestCase):
         cls.genotypePhenotype = cls.phenotypeAssociationSet.getAssociations(
             request=None, featureSets=cls.featureSets)[0]
         cls.genotypePhenotypeId = cls.genotypePhenotype.id
-        cls.rnaQuantificationSet = cls.dataset.getRnaQuantificationSets()[0]
-        cls.rnaQuantificationSetId = cls.rnaQuantificationSet.getId()
-        cls.rnaQuantification = cls.rnaQuantificationSet.getRnaQuantifications(
-            )[0]
-        cls.rnaQuantificationId = cls.rnaQuantification.getId()
-        cls.expressionLevel = cls.rnaQuantification.getExpressionLevels(
-            1, 2)[0]
-        cls.expressionLevelId = cls.expressionLevel.getId()
+        # cls.rnaQuantificationSet = cls.dataset.getRnaQuantificationSets()[0]
+        # cls.rnaQuantificationSetId = cls.rnaQuantificationSet.getId()
+        # cls.rnaQuantification =
+        #    cls.rnaQuantificationSet.getRnaQuantifications()[0]
+        # cls.rnaQuantificationId = cls.rnaQuantification.getId()
+        # cls.expressionLevel = cls.rnaQuantification.getExpressionLevels(
+        #     1, 2)[0]
+        # cls.expressionLevelId = cls.expressionLevel.getId()
 
     def sendPostRequest(self, path, request):
         """
@@ -457,18 +457,21 @@ class TestFrontend(unittest.TestCase):
             protocol.Variant,
             self.variantId)
 
+    @unittest.skip("Disabled")
     def testGetExpressionLevel(self):
         self.getObjectTest(
             self.sendGetExpressionLevel,
             protocol.ExpressionLevel,
             self.expressionLevelId)
 
+    @unittest.skip("Disabled")
     def testGetRnaQuantification(self):
         self.getObjectTest(
             self.sendGetRnaQuantification,
             protocol.RnaQuantification,
             self.rnaQuantificationId)
 
+    @unittest.skip("Disabled")
     def testGetRnaQuantificationSet(self):
         self.getObjectTest(
             self.sendGetRnaQuantificationSet,
@@ -505,6 +508,7 @@ class TestFrontend(unittest.TestCase):
             "associations",
             self.genotypePhenotypeId)
 
+    @unittest.skip("Disabled")
     def testExpressionLevelsSearch(self):
         self.searchObjectTest(
             self.sendExpressionLevelsSearch,
@@ -512,6 +516,7 @@ class TestFrontend(unittest.TestCase):
             "expression_levels",
             self.expressionLevelId)
 
+    @unittest.skip("Disabled")
     def testRnaQuantificationsSearch(self):
         self.searchObjectTest(
             self.sendRnaQuantificationsSearch,
@@ -519,6 +524,7 @@ class TestFrontend(unittest.TestCase):
             "rna_quantifications",
             self.rnaQuantificationId)
 
+    @unittest.skip("Disabled")
     def testRnaQuantificationSetsSearch(self):
         self.searchObjectTest(
             self.sendRnaQuantificationSetsSearch,
